Resolved compiler warning.
git-svn-id: https://svn.aps.anl.gov/AOP/oag/trunk/apps/src/elegant@27135 c1b61890-85f3-ea29-e4d6-a272fe79f168

+++ sdds_beam.c
@@ -754,7 +754,7 @@ long new_sdds_beam(
       sprintf(buffer, "%ld", (long)beam->particle[i][bunchIndex]);
       hadd(hashTable, buffer, strlen(buffer), NULL);
     }
-    printf("%d bunches present\n", (long)hcount(hashTable));
+    printf("%d bunches present\n", (int)hcount(hashTable));
 #if DEBUG
     printf("id:[%ld, %ld], b:[%ld, %ld]\n", (long)hcount(hashTable), minID, maxID,
            minBunch, maxBunch);
